+++ lib/src/resource_invoker.dart
@@ -1,0 +1,17 @@
+//Copyright (C) 2013 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+part of rest_server;
+
+/**
+ * Class helps invoke [Resource].
+ */
+class ResourceInvoker {
+  
+  /**
+   * Invoke [resource] with [request] info and return result of invocation.
+   */
+  dynamic invoke(HttpRequest request, Resource resource) {
+    throw new Exception("Not implemented yet");
+  }
+}